backend/media-service/requirements-full.txt
# 核心Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
gunicorn>=21.2.0
python-multipart>=0.0.6

# 基础配置
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP客户端
requests>=2.31.0
httpx>=0.25.0

# Azure核心服务 (最小集)
azure-cognitiveservices-speech>=1.30.0
azure-storage-blob>=12.0.0
openai>=1.0.0

# 基础工具
python-dateutil>=2.8.2
